+++ app/src/userpages/components/StreamPage/Edit/ConfigureView/configureView.pcss
