+++ app/src/userpages/components/StreamPage/Edit/SecurityView/Slider.jsx
